requirements.txt
langchain>=0.0.27
openai>=0.27.0
python-dotenv>=0.19.0
requests>=2.28.0
numpy>=1.21.0
pandas>=1.3.0
swarm @ git+https://github.com/openai/swarm.git